Rename module to enable pytest 7.0.0
