+++ tests/unit/summaries/metrics/test_mean_absolute_error.py
@@ -9,7 +9,7 @@ import xarray as xr
 from pywatts.core.exceptions.input_not_available import InputNotAvailable
 from pywatts.core.exceptions.invalid_input_exception import InvalidInputException
 from pywatts.summaries.mae_summary import MAE
-from tests.unit.summaries.metrics.base_test_metric_base import BaseTestMetricBase
+from tests.unit.summaries.metrics.test_base_metric_base import BaseTestMetricBase
 
 
 class TestMAE(BaseTestMetricBase, unittest.TestCase):
