+++ tests/unit/summaries/metrics/test_root_mean_squared_error.py
@@ -7,7 +7,7 @@ import xarray as xr
 
 from pywatts.core.exceptions.input_not_available import InputNotAvailable
 from pywatts.summaries import RMSE
-from tests.unit.summaries.metrics.base_test_metric_base import BaseTestMetricBase
+from tests.unit.summaries.metrics.test_base_metric_base import BaseTestMetricBase
 
 
 class TestRMSE(BaseTestMetricBase, unittest.TestCase):
